manual implementation for message
