+++ agreement/player.go
@@ -392,7 +392,7 @@ func (p *player) enterRound(r routerHandle, source event, target round) []action
 
 	if e.t() == payloadPipelined {
 		e := e.(payloadProcessedEvent)
-		msg := message{messageHandle: 0, Tag: protocol.ProposalPayloadTag, UnauthenticatedProposal: e.UnauthenticatedPayload} // TODO do we want to keep around the original handle?
+		msg := message{MessageHandle: 0, Tag: protocol.ProposalPayloadTag, UnauthenticatedProposal: e.UnauthenticatedPayload} // TODO do we want to keep around the original handle?
 		a := verifyPayloadAction(messageEvent{T: payloadPresent, Input: msg}, p.Round, e.Period, e.Pinned)
 		actions = append(actions, a)
 	}
@@ -571,7 +571,7 @@ func (p *player) handleMessageEvent(r routerHandle, e messageEvent) (actions []a
 		}
 
 		// relay as the proposer
-		if e.Input.messageHandle == nil {
+		if e.Input.MessageHandle == nil {
 			var uv unauthenticatedVote
 			switch ef.t() {
 			case payloadPipelined, payloadAccepted:
